pc_backend/requirements.txt
# PC Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
faster-whisper>=0.10.0
google-generativeai>=0.8.0
numpy>=1.24.0
piper-tts>=1.2.0
python-dotenv>=1.0.0
soundfile>=0.12.0
